% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status.R
\name{check_status}
\alias{check_status}
\title{Identifies open nomenclature in species scientific name and classification under species level}
\usage{
check_status(scientificName = NULL)
}
\arguments{
\item{scientificName}{species name in a raw format}
}
\description{
This function identifies aff., cf. and subsp. var. in species names. It creates a new column with the original string and the new suggested name.
}
\examples{
check_status("Lindsaea lancea var. falcata")
check_status(c("Lindsaea lancea var. falcata", "Asplenium Aff. truncorum"))

}
\author{
Sara Mortara
}
