% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGADM.R
\name{getGADM}
\alias{getGADM}
\title{Function to download GADM data for a country at a specific level}
\usage{
getGADM(cod, level = 4, type = "sf", best = FALSE, destfolder = "./GADM/", ...)
}
\arguments{
\item{cod}{Three letter ISO code for the country.}

\item{level}{Administrative level. 0 stands for country and 1 to 4 are the
successive successive nested levels. Not every country has the same depth and
the nestedness is sequential. This means a country with two levels has levels
0 and 1, not 0 and 4, and that units such as states or municipalities can
correspond to different levels depending on the country}

\item{type}{The class of spatial object to be downloaded. Defaults to sf
but can be set to sp}

\item{best}{TRUE Checks if finer resolutions are available and avoids
downloading data at coarser resolutions. Defaults to TRUE (will not download
coarser resolution files)}

\item{destfolder}{The destination folder}

\item{...}{Options for download.file()}
}
\description{
This function downloads the GADM shapefile for a country. If the file is
already on disk it will not perform the download. The function can check if
better resolutions are available in disk.
}
\details{
https://gadm.org/data.html
}
\examples{
\dontrun{
getGADM(cod = "BRA", level = 3)
getGADM(cod = "BRA", level = 1, best = TRUE) # will not download if lower levels exist
getGADM(cod = "BRA", level = 3, type = "sp")
}
}
