% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGAZ.R
\name{getGAZ}
\alias{getGAZ}
\title{Function to download DIVA-GIS gazetteer data for a country}
\usage{
getGAZ(cod, destfolder = "GAZ", unzip = TRUE, ...)
}
\arguments{
\item{cod}{Three letter ISO code for the country.}

\item{destfolder}{The destination folder}

\item{unzip}{Unzip the downloaded files? Defaults to TRUE}

\item{...}{Options for download.file()}
}
\description{
This function downloads the DIVA-GIS gazetteer data for a country.
If the file is already on disk it will not perform the download.
}
\details{
https://gadm.org/data.html
}
\examples{
\dontrun{
getGAZ(cod = "COL", unzip = TRUE)
countries <- c("COL", "BRA", "ECU", "PER", "VEN", "BOL")
lapply(countries, getGAZ)
}

}
\author{
Andrea Sánchez-Tapia & Sara Mortara
}
