% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWDPA.R
\name{getWDPA}
\alias{getWDPA}
\title{Function to download WDPA data for a country}
\usage{
getWDPA(cod, destfolder = "WDPA", unzip = TRUE, ...)
}
\arguments{
\item{cod}{Three letter ISO code for the country.}

\item{destfolder}{The destination folder}

\item{unzip}{Unzip the downloaded files? Defaults to TRUE}

\item{...}{Options for download.file()}
}
\description{
This function downloads the WDPA shapefile for a country (www.protectedplanet.net).
If the file is already on disk it will not perform the download.
}
\details{
https://www.protectedplanet.net
}
\examples{
\dontrun{
getWDPA(cod = "BRA", unzip = TRUE)
}

}
\author{
Andrea Sánchez-Tapia & Sara Mortara
}
