% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocc.R
\docType{package}
\name{rocc}
\alias{rocc}
\title{rocc: Workflows for Downloading Occurrence Data from speciesLink, FB2020, and JABOT API}
\description{
\code{rocc} checks, formats and downloads occurrence data from various sources
}
\section{functions inside this package}{
NA
}

