% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_flora.R
\name{search_flora}
\alias{search_flora}
\title{Returns species scientific name lists from phytogeographic domain,
state and/or endemism status from the List of Species of the Brazilian Flora
2020 database}
\usage{
search_flora(
  domain = c("Cerrado", "Mata Atlântica"),
  states = "MG",
  endemism = NULL
)
}
\arguments{
\item{domain}{a domain in Mata Atlântica, Cerrado, Pantanal, Pampa, Amazônia}

\item{states}{Brazilian states, capitalized (e.g. "MA")}

\item{endemism}{logical, return species that are endemic or not from Brazil
Defaults to NULL to return all species}
}
\value{
a list of taxon IDs
}
\description{
Returns species scientific name lists from phytogeographic domain,
state and/or endemism status from the List of Species of the Brazilian Flora
2020 database
}
\examples{
# All endemic species from the caatinga in Pernambuco
search_flora(domain = "Caatinga",
             state = "PE",
             endemism = TRUE)


}
