% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_infraspecies.R
\name{set_infraspecies}
\alias{set_infraspecies}
\title{Adds infraspecies identification if missing}
\usage{
set_infraspecies(subspecies, type)
}
\arguments{
\item{subspecies}{Species name containing three stings: Genus, species epithet, and infraspecies epithet}

\item{type}{Type of infraspecies classification. Any type in: `"var"` or `"subsp"` or `"form"`.}
}
\description{
Adds infraspecies identification if missing
}
\examples{
set_infraspecies("Lindsaea lancea lancea", type = "var")

}
