% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spLink_url.R
\name{spLink_url}
\alias{spLink_url}
\title{Gets occurrence data from speciesLink}
\usage{
spLink_url(dir = "results/", filename = "output",
  scientificname = NULL, collectioncode = NULL, County = NULL,
  Coordinates = NULL, coordinatesQuality = NULL, MaxRecords = NULL,
  Model = NULL, Scope = NULL, Synonyms = "no synomyms",
  Images = NULL)
}
\arguments{
\item{dir}{Path to directory where the file will be saved. Default is to create a "results/" directory}

\item{filename}{Name of the output file}

\item{scientificname}{Genus and epithet separated by space}

\item{collectioncode}{Any collection available at speciesLink. Example: ALCB, E, INPA, MOBOT_BR}

\item{County}{Any municipality name. No ASCII characters allowed}

\item{Coordinates}{Specify if records should have coordinates. Default is "no check" but it also accepts "Yes", No", "Original", "Automatic", "Blocked" or "no check"}

\item{coordinatesQuality}{Any character in "Good" or "Bad" to select specific type of coordinates}

\item{MaxRecords}{Numeric. Maximum number of records to be required}

\item{Model}{If file should be prepared in "DwC" or "modelling" format}

\item{Scope}{Group to be required. If NULL searches all groups. Any in "plants", "animals", "microrganisms" or "fossils"}

\item{Synonyms}{If species names should be checked for synonyms in a specific dictionary. Set to "species2000" for search in Catálogo da Vida species2000, "flora2020" for Flora do Brasil 2020, "MycoBank" for MycoBank, "AlgaeBase" for AlgaeBase, "DSMZ" for  DSMZ Prokaryotic Nomenclature Up-to-Date, "Moure" for Catálogo de Abelhas Moure or "no synonyms".}

\item{Images}{If select only records with images. Default is NULL. It accepts: "Yes", "Live", "Polen", "Wood"}
}
\value{
A list of two elemnts. The first element is a character string containing the url search and the second element is a data.frame with the search result
}
\description{
This function access version beta 0.1 of speciesLink API and returns occurrence data from species.
}
\examples{

ex01 <- spLink_url(filename = "ex01",
scientificname =  c("Eugenia platyphylla", "Chaetocalyx acutifolia"),
Scope="plants")

}
\author{
Sara Mortara
}
