% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_bfg.R
\name{suggest_bfg}
\alias{suggest_bfg}
\title{Function suggest a valid species name for Brazilian plants}
\usage{
suggest_bfg(scientificName)
}
\arguments{
\item{scientificName}{Character. A species scientific name without authors, ideally already passed by the string check in `rocc::check_status`}
}
\value{
A data frame with the name provided by the user and a new suggested name
}
\description{
Function suggest a valid species name for Brazilian plants
}
