% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_flora.R
\name{update_flora}
\alias{update_flora}
\title{Downloads, updates and formats the data from the ipt server of the List of Species of the Brazilian Flora}
\usage{
update_flora(force_update = FALSE)
}
\arguments{
\item{force_update}{Logical. Forces update in case there is already an ipt file in cache}
}
\description{
Downloads, updates and formats the data from the ipt server of the List of Species of the Brazilian Flora
}
