% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flora.R
\name{check_flora}
\alias{check_flora}
\title{Function to check species scientific name in Brazilian Flora 2020 database}
\usage{
check_flora(scientificName, get_synonyms = TRUE, infraspecies = FALSE)
}
\arguments{
\item{scientificName}{Character. A species scientific name without authors, ideally already passed by the string check in `rocc::check_status`. Accepts only one name per time. Use `lapply` ou functions from `purr` package to run for multiple species}

\item{get_synonyms}{Logical. If `get_synonyms = TRUE` (default) returns a second element containing information of all synonyms a species has}

\item{infraspecies}{Logical. If `infraspecies = TRUE` returns accepted name of any infraspecies classification}
}
\value{
A list with one or two elements. If `get_synonyms = TRUE` the second element of the list contains the names and additional information of synonyms or a NULL element if the species has no synonyms
}
\description{
Function to check species scientific name in Brazilian Flora 2020 database
}
\examples{
# single species
check_flora("Dalbergia nigra")

# more than one species w/ lapply from base
lapply(c("Dalbergia nigra", "Aspidosperma discolor"),
       check_flora)

}
