% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status.R
\name{check_status}
\alias{check_status}
\title{Identifies open nomenclature in species scientific name and classification under species level}
\usage{
check_status(scientificName = NULL)
}
\arguments{
\item{scientificName}{species name in a raw format}
}
\value{
Data frame with `scientificName` as in the original input, `scientificName_status` with the flags in original data and `scientificName_new` with a suggestion for a more correct species name. The column `scientificName_status` accepts:
\describe{
\item{\code{possibly_ok}}{scientific name following the expected pattern Genus epithet}
\item{\code{not_Genus_epithet_format}}{scientific name not following the expected pattern Genus epithet}
\item{\code{variety}}{species scientific name with variety}
\item{\code{subspecies}}{species scientific name with subspecies}
\item{\code{conferre}}{open nomenclature cf. in species scientific name}
\item{\code{affinis}}{open nomenclature aff. in species scientific name}
\item{\code{name_w_authors}}{species scientific name has authors}
\item{\code{not_name_has_digits}}{species scientific name has digits, not a valid name}
\item{\code{indet}}{species identified only at genus level}
\item{\code{family_as_genus}}{species family as genus, not a valid name}
}
}
\description{
This function identifies aff., cf. and subsp. var. in species names. It creates a new column with the original string and the new suggested name. It also flags problematic names (character string with numbers, authors, wrong case, or other names besides genus and epithet).
}
\details{
Flags returned by column
}
\examples{
check_status("Lindsaea lancea var. falcata")
check_status(c("Lindsaea lancea var. falcata", "Asplenium Aff. truncorum"))
check_status("Asplenium sp.")

}
\author{
Sara Mortara
}
