% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taxon.R
\name{check_taxon}
\alias{check_taxon}
\title{Function to check species scientific name in Brazilian Flora 2020 database}
\usage{
check_taxon(scientificName)
}
\arguments{
\item{scientificName}{a species scientific name without authors, ideally already passed by the string check in `rocc::check_status`. Accepts only one name per time. Use `lapply` ou functions from `purr` package to run for multiple species.}
}
\value{
A list with one or two elements. If a species has synonyms the second element of the list contains the names and additional information of synonyms.
}
\description{
Function to check species scientific name in Brazilian Flora 2020 database
}
\examples{
# single species
check_taxon("Dalbergia nigra")

# more than one species w/ lapply from base
lapply(c("Darbergia nigra", "Aspidosperma discolor"),
       check_taxon)

}
