% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_flora.R
\name{search_flora}
\alias{search_flora}
\title{Searches in the List of Species of the Brazilian Flora 2020 database}
\usage{
search_flora(
  domain = NULL,
  stateProvince = NULL,
  endemism = NULL,
  lifeform = NULL,
  habitat = NULL,
  force_update = FALSE
)
}
\arguments{
\item{domain}{a domain in Mata Atlântica, Cerrado, Pantanal, Pampa, Amazônia}

\item{stateProvince}{Brazilian states, capitalized (e.g. "MA")}

\item{endemism}{logical, return species that are endemic or not from Brazil
Defaults to NULL to return all species}

\item{lifeform}{character, search species with the following life forms:
\describe{
  \item{\code{Arbusto}}{}
  \item{\code{Árvore}}{}
  \item{\code{Bambu}}{}
  \item{\code{Dracenoide}}{}
  \item{\code{Erva}}{}
  \item{\code{Liana/volúvel/trepadeira}}{}
  \item{\code{Palmeira}}{}
  \item{\code{Subarbusto}}{}
  \item{\code{Suculenta}}{}
  \item{\code{Desconhecida}}{}
}}

\item{habitat}{character, search species in the following habitat options:
\describe{
  \item{\code{Aquática}}{}
  \item{\code{Epífita}}{}
  \item{\code{Hemiepífita}}{}
  \item{\code{Hemiparasita}}{}
  \item{\code{Parasita}}{}
  \item{\code{Rupícola}}{}
  \item{\code{Saprófita}}{}
  \item{\code{Terrícola}}{}
  \item{\code{Desconhecido}}{}
}}

\item{force_update}{Logical. Update the flora IPT database in cache?}
}
\value{
a data frame with taxon IDs, scientific name and binomial name
without authors
}
\description{
Returns species scientific name lists from phytogeographic domain,
state and/or endemism status from the List of Species of the Brazilian Flora
2020 database
}
\examples{
# All endemic species from the caatinga in Pernambuco
search_flora(domain = "Caatinga",
             state = "PE",
             endemism = TRUE)


}
