% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_flora.R
\name{search_flora}
\alias{search_flora}
\title{Searches in the List of Species of the Brazilian Flora 2020 database}
\usage{
search_flora(
  domain = NULL,
  stateProvince = NULL,
  life_form = NULL,
  habitat = NULL,
  vegetation_type = NULL,
  endemism = NULL,
  force_update = FALSE
)
}
\arguments{
\item{domain}{a domain in Mata Atlântica, Cerrado, Pantanal, Pampa, Amazônia}

\item{stateProvince}{Two-letter code for Brazilian states (e.g. "MA")}

\item{life_form}{character, search species with the following life forms}

\item{habitat}{character, search species in the habitat options according to the FB2020}

\item{vegetation_type}{character, filter by vegetation type according to the FB2020}

\item{endemism}{logical, return species that are endemic or not from Brazil
Defaults to NULL to return all species}

\item{force_update}{Logical. Update the flora IPT database in cache?}
}
\value{
a data frame with taxon IDs, scientific name and binomial name
without authors
}
\description{
Returns species scientific name lists from phytogeographic domain,
state, life form and/or endemism status from the List of Species of the Brazilian Flora
2020 database
}
\examples{
# All endemic species from the caatinga in Pernambuco
\dontrun{
search_flora(domain = "Caatinga",
             state = "PE",
             endemism = TRUE)
}


}
