% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_flora.R
\name{suggest_flora}
\alias{suggest_flora}
\title{Function suggest a valid species name for Brazilian plants}
\usage{
suggest_flora(species)
}
\arguments{
\item{species}{Character. A species scientific name without authors, ideally already passed by the string check in `Rocc::check_status()`}
}
\value{
A data frame with the name provided by the user and a new suggested name
}
\description{
Function suggest a valid species name for Brazilian plants
}
\author{
Sara Mortara & Andrea Sánchez-Tapia
}
